+++ static/components/f-date.js
@@ -6,5 +6,17 @@ customElements.define(
     get inputAttributes() {
       return { type: "date", class: "input is-narrow" };
     }
+
+    get value() {
+      let value = super.value;
+
+      if (typeof value === "string") {
+        value = new Date(value);
+        super.value = value;
+      }
+
+      // Get the value in the format "YYYY-MM-DD"
+      return value.toISOString().split("T")[0];
+    }
   },
 );