+++ CHANGELOG.md
@@ -4,7 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [0.2.0] - Unreleased
+### Changed
+- Rename `/src` to `/core`.
+- Moved `/storage` to the root folder.
+- Moved `/src/routes/templates` to `/core/templates`.
+- Moved `/src/types.ts` to the root folder.
+
+### Fixed
+- Favicon error.
+- `f-date` field.
+
 ## [0.1.0] - 2024-02-08
 First version
 
+[0.2.0]: https://github.com/lumeland/cms/compare/v0.1.0...HEAD
 [0.1.0]: https://github.com/lumeland/cms/releases/tag/v0.1.0